@@ -2,7 +2,7 @@
 
 	<PropertyGroup>
 		<OutputType>Exe</OutputType>
-		<TargetFramework>net6.0</TargetFramework>
+		<TargetFramework>net9.0</TargetFramework>
 		<LangVersion>9.0</LangVersion>
 		<ImplicitUsings>enable</ImplicitUsings>
 		<Nullable>enable</Nullable>
@@ -15,7 +15,7 @@
 		<PackageReference Include="K4os.Compression.LZ4" Version="1.3.5" />
 		<PackageReference Include="K4os.Compression.LZ4.Streams" Version="1.3.5" />
 		<PackageReference Include="ZstdNet" Version="1.4.5" />
-		<ProjectReference Include="..\..\src\NativeCompressions.Lz4\NativeCompressions.Lz4.csproj" />
+		<ProjectReference Include="..\..\src\NativeCompressions.LZ4\NativeCompressions.LZ4.csproj" />
 		<ProjectReference Include="..\..\src\NativeCompressions.ZStandard\NativeCompressions.ZStandard.csproj" />
 	</ItemGroup>
 
